refactor: extract Function abstraction and split domain/api types
