+++ packages/shared/lib/services/on-event-scripts.service.ts
@@ -3,32 +3,13 @@ import { env } from '@nangohq/utils';
 
 import configService from './config.service.js';
 import remoteFileService from './file/remote.service.js';
+import { eventTypeMapper } from './functions/mappers.js';
 import { resolveLocalFileName } from '../utils/utils.js';
 import { increment } from './sync/config/config.service.js';
 
 import type { DBEnvironment, DBOnEventScript, DBTeam, OnEventScript, OnEventScriptsByProvider, OnEventType } from '@nangohq/types';
 
 const TABLE = 'on_event_scripts';
-
-const EVENT_TYPE_MAPPINGS: Record<DBOnEventScript['event'], OnEventType> = {
-    POST_CONNECTION_CREATION: 'post-connection-creation',
-    PRE_CONNECTION_DELETION: 'pre-connection-deletion',
-    VALIDATE_CONNECTION: 'validate-connection'
-} as const;
-
-const eventTypeMapper = {
-    fromDb: (event: DBOnEventScript['event']): OnEventType => {
-        return EVENT_TYPE_MAPPINGS[event];
-    },
-    toDb: (eventType: OnEventType): DBOnEventScript['event'] => {
-        for (const [key, value] of Object.entries(EVENT_TYPE_MAPPINGS)) {
-            if (value === eventType) {
-                return key as DBOnEventScript['event'];
-            }
-        }
-        throw new Error(`Unknown event type: ${eventType}`); // This should never happen
-    }
-};
 
 const dbMapper = {
     to: (script: OnEventScript): DBOnEventScript => {
