+++ packages/types/lib/index.ts
@@ -99,6 +99,7 @@ export type * from './checkpoint/db.js';
 
 export type * from './mcp/api.js';
 export type * from './functions/api.js';
+export type * from './functions/domain.js';
 
 export type * from './lambda/index.js';
 export type * from './authz/types.js';
